requirements.txt
google-generativeai==0.8.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pytesseract==0.3.13
python-dotenv==1.1.1
streamlit==1.50.0